Move Studio panel UI to its own file.
